OJHelper/requirements.txt
customtkinter
selenium
google-generativeai
openai
anthropic
requests
packaging